+++ lib/mix/tasks/mob.mutate.ex
@@ -60,6 +60,14 @@ defmodule Mix.Tasks.Mob.Mutate do
 
   @impl Mix.Task
   def run(args) do
+    if MobDev.TaskHelp.help_requested?(args) do
+      MobDev.TaskHelp.print_module_help(__MODULE__)
+    else
+      do_run(args)
+    end
+  end
+
+  defp do_run(args) do
     {opts, _argv, invalid} = OptionParser.parse(args, strict: @switches)
 
     unless invalid == [] do